refactor(Account): turned off newline-after-import + moved imports
